feat: document render module
